==> 3.FineTuningAnnotations.Rmd <==
---
title: "Tuning SingleR Annotations"
output: html_notebook
---

Here we will look at trying to improve some of the annotations from SingleR and exploring that part of the process a little more.

Firstly lets get the libraries that we are going to need imported.

```{r libraries}
library(tidyverse)
library(Seurat)
library(SingleR)
library(BiocParallel)
library(plotly)
```

## Loading in Data

Now we can grab the final SeuratObject that was created in the initial exploration of the data.

```{r}
manley_data = readRDS("Manley_SMI/results/rds/final_explore.rds")
```

Lets first start by plotting the annotations again to see the cell types that we were able to assign.

```{r, fig.height = 10, fig.width = 10}
DimPlot(manley_data, label = T, group.by = "SingleR.labels", repel = T) +
  theme(legend.position = 'none')
```

## T Cells

Lets look deeper at the T-cell and the NK cells to see if they are fairly concentrated or if they are dispersed.

```{r, fig.height = 10, fig.width = 10}
Idents(manley_data) = "SingleR.labels"
DimPlot(manley_data, label = T, group.by = "SingleR.labels", 
        cells.highlight = list(WhichCells(manley_data, idents = c("CD4 T cell",
                                                                  "CD8 T cell",
                                                                  "NKT cell",
                                                                  "NK cell"))), 
        repel = T,
        raster.dpi = c(512, 512))
```

```{r, fig.height = 10, fig.width = 10}
Idents(manley_data) = "seurat_clusters"
DimPlot(manley_data, label = T, group.by = "seurat_clusters", 
        cells.highlight = list(WhichCells(manley_data, idents = c("7",
                                                                  "5",
                                                                  "29"))), 
        repel = T,
        raster.dpi = c(512, 512))
```

```{r}
t_cell_data = subset(x = manley_data, subset = SingleR.labels %in% c("CD8 T cell",
                                                                   "CD4 T cell",
                                                                   "NKT cell",
                                                                   "NK cell"))
Idents(t_cell_data) = "slide_fov"
```

```{r}
#t_cell_data = RunPCA(object = t_cell_data, assay = "SCT", npcs = 200, reduction.name = "pca_t_cells")
#saveRDS(t_cell_data, "Manley_SMI/results/rds/t_cell_seurat_pca.rds")
# t_cell_data = readRDS("Manley_SMI/results/rds/t_cell_seurat_pca.rds")
# t_cell_data = RunUMAP(object = t_cell_data, reduction = "pca_t_cells", dims = 1:50, assay = "SCT",
#                       n.components = 2, reduction.name = "SCT_umap_t_cells")
# #clustering with louvain
# t_cell_data = FindNeighbors(t_cell_data, dims = 1:50, verbose = TRUE, reduction = "pca_t_cells")
# t_cell_data = FindClusters(t_cell_data, verbose = TRUE, resolution = 0.4)
# 
# saveRDS(t_cell_data, "Manley_SMI/results/rds/T-cell_Louvain_Clustering.rds")
t_cell_data = readRDS("Manley_SMI/results/rds/T-cell_Louvain_Clustering.rds")
```

```{r}
DimPlot(t_cell_data, reduction = "SCT_umap_t_cells", group.by = "seurat_clusters", label = T)
```

```{r}
#DEGs = FindAllMarkers(t_cell_data, assay = "SCT")
#saveRDS(DEGs, "Manley_SMI/results/rds/T-cell_DEGS_inLouvainClusters.rds")
DEGs = readRDS("Manley_SMI/results/rds/T-cell_DEGS_inLouvainClusters.rds")

```

```{r}
t_cell_data$Manual.Phenotype = NA
t_cell_data$Manual.Phenotype = ifelse(is.na(t_cell_data$Manual.Phenotype) & t_cell_data$seurat_clusters == 0, "CD8 T cell", t_cell_data$Manual.Phenotype)
t_cell_data$Manual.Phenotype = ifelse(is.na(t_cell_data$Manual.Phenotype) & t_cell_data$seurat_clusters == 1, "CD4 T cell", t_cell_data$Manual.Phenotype)
t_cell_data$Manual.Phenotype = ifelse(is.na(t_cell_data$Manual.Phenotype) & t_cell_data$seurat_clusters == 2, "Classical Monocyte", t_cell_data$Manual.Phenotype)
t_cell_data$Manual.Phenotype = ifelse(is.na(t_cell_data$Manual.Phenotype) & t_cell_data$seurat_clusters == 3, "NKT cell", t_cell_data$Manual.Phenotype)
t_cell_data$Manual.Phenotype = ifelse(is.na(t_cell_data$Manual.Phenotype) & t_cell_data$seurat_clusters == 4, "B cell", t_cell_data$Manual.Phenotype)
t_cell_data$Manual.Phenotype = ifelse(is.na(t_cell_data$Manual.Phenotype) & t_cell_data$seurat_clusters == 5, "Collecting Duct cells", t_cell_data$Manual.Phenotype)
t_cell_data$Manual.Phenotype = ifelse(is.na(t_cell_data$Manual.Phenotype) & t_cell_data$seurat_clusters == 6, "Possibly Mid-Replication", t_cell_data$Manual.Phenotype)
```

```{r}
DimPlot(t_cell_data, reduction = "SCT_umap_t_cells", group.by = "Manual.Phenotype", label = T)
```

Number of "T-cells" (annotated originally by SingleR as T-cell) across the different slides: 
```{r}
meta = t_cell_data@meta.data
meta %>% group_by(tissue, fov) %>% summarise(cells = n(), .groups = "keep") %>% spread(tissue, cells) %>% colSums(na.rm= T)
```

Looks like there is about half the number of cells on the RCC5 slide. Lets look at the overall data to see if there is the same trend or just less T-cells being annotated on the last slide.

```{r}
manley_data@meta.data %>%
  group_by(tissue, fov) %>%
  summarise(cells = n(), .groups = "keep") %>%
  spread(tissue, cells) %>%
  colSums(na.rm=T)
```

There is a lower number of cells in total on the last RCC5 slide, but not quite the extreme that the annotated t-cells are. Will be something to keep in mind as moving forward.

Lets add the manually annotated T-cell labels back to the `manley_data`.

```{r, fig.height = 10, fig.width = 10}
manley_data$final_labels = manley_data$SingleR.labels
#meta$orig.ident
manley_data$final_labels[manley_data$cell_ID %in% meta$cell_ID] = meta$Manual.Phenotype
DimPlot(manley_data, group.by = "final_labels", label = T, repel = T)  +
  theme(legend.position = 'none')
```
# Macrophage

The macrophage markers that are included in the SMI data are CD163 and CD86 (cross referenced with the M1/M2 markers included on the TAM mIF panel for Tworoger's data).

```{r, fig.height=10, fig.width=10}
FeaturePlot(manley_data, features = c("CD163", "CD86", "MRC1"))
```

We can check to see if the MNP cell types are assigned to cells in the same area.

```{r, fig.height = 10, fig.width = 10}
Idents(manley_data) = "final_labels"
DimPlot(manley_data, label = T, group.by = "final_labels", 
        cells.highlight = list(WhichCells(manley_data, idents = c("MNP-c/dendritic cell",
                                                                  "MNP-d/Tissue macrophage",
                                                                  "MNP-b/non-classical monocyte derived",
                                                                  "MNP-a/classical monocyte derived"))), 
        repel = T,
        raster.dpi = c(512, 512))
```

```{r}
MNP_cell_data = subset(x = manley_data, subset = SingleR.labels %in% c("MNP-c/dendritic cell",
                                                                  "MNP-d/Tissue macrophage",
                                                                  "MNP-b/non-classical monocyte derived",
                                                                  "MNP-a/classical monocyte derived"))
Idents(MNP_cell_data) = "slide_fov"
```

```{r}
# MNP_cell_data = RunPCA(object = MNP_cell_data, assay = "SCT", npcs = 200, reduction.name = "pca_mnp_cells")
# saveRDS(MNP_cell_data, "Manley_SMI/results/rds/mnp_cells_seurat_pca.rds")
# MNP_cell_data = readRDS("Manley_SMI/results/rds/mnp_cells_seurat_pca.rds")
# MNP_cell_data = RunUMAP(object = MNP_cell_data, reduction = "pca_mnp_cells", dims = 1:50, assay = "SCT",
#                       n.components = 2, reduction.name = "SCT_umap_mnp_cells")
# #clustering with louvain
# MNP_cell_data = FindNeighbors(MNP_cell_data, dims = 1:50, verbose = TRUE, reduction = "pca_mnp_cells")
# MNP_cell_data = FindClusters(MNP_cell_data, verbose = TRUE, resolution = 0.4)
# 
# saveRDS(MNP_cell_data, "Manley_SMI/results/rds/MNP_cells_Louvain_Clustering.rds")
MNP_cell_data = readRDS("Manley_SMI/results/rds/MNP_cells_Louvain_Clustering.rds")
```

```{r}
DimPlot(MNP_cell_data, reduction = "SCT_umap_mnp_cells", group.by = "seurat_clusters", label = T)
```

```{r}
#DEGs = FindAllMarkers(MNP_cell_data, assay = "SCT")
#saveRDS(DEGs, "Manley_SMI/results/rds/MNP_cells_DEGS_inLouvainClusters.rds")
DEGs = readRDS("Manley_SMI/results/rds/MNP_cells_DEGS_inLouvainClusters.rds")
DEGs %>%
  filter(p_val_adj < 0.01 & abs(avg_log2FC) > 0.5)

```

```{r}
MNP_cell_data$Manual.Phenotype = NA
MNP_cell_data$Manual.Phenotype = ifelse(is.na(MNP_cell_data$Manual.Phenotype) & MNP_cell_data$seurat_clusters == 0, "Intermediate Monocyte", MNP_cell_data$Manual.Phenotype)
MNP_cell_data$Manual.Phenotype = ifelse(is.na(MNP_cell_data$Manual.Phenotype) & MNP_cell_data$seurat_clusters == 1, "CD8 T cell", MNP_cell_data$Manual.Phenotype)
MNP_cell_data$Manual.Phenotype = ifelse(is.na(MNP_cell_data$Manual.Phenotype) & MNP_cell_data$seurat_clusters == 2, "Classic Monocyte/Macrophage", MNP_cell_data$Manual.Phenotype)
MNP_cell_data$Manual.Phenotype = ifelse(is.na(MNP_cell_data$Manual.Phenotype) & MNP_cell_data$seurat_clusters == 3, "Collecting Duct cells", MNP_cell_data$Manual.Phenotype)
MNP_cell_data$Manual.Phenotype = ifelse(is.na(MNP_cell_data$Manual.Phenotype) & MNP_cell_data$seurat_clusters == 4, "CD163+ Macrophage", MNP_cell_data$Manual.Phenotype)
MNP_cell_data$Manual.Phenotype = ifelse(is.na(MNP_cell_data$Manual.Phenotype) & MNP_cell_data$seurat_clusters == 5, "Collecting Duct cells", MNP_cell_data$Manual.Phenotype)
MNP_cell_data$Manual.Phenotype = ifelse(is.na(MNP_cell_data$Manual.Phenotype) & MNP_cell_data$seurat_clusters == 6, "Non-Classical Monocyte", MNP_cell_data$Manual.Phenotype)
MNP_cell_data$Manual.Phenotype = ifelse(is.na(MNP_cell_data$Manual.Phenotype) & MNP_cell_data$seurat_clusters == 7, "Myeloid DC", MNP_cell_data$Manual.Phenotype)
MNP_cell_data$Manual.Phenotype = ifelse(is.na(MNP_cell_data$Manual.Phenotype) & MNP_cell_data$seurat_clusters == 8, "Intermediate Monocyte", MNP_cell_data$Manual.Phenotype)
MNP_cell_data$Manual.Phenotype = ifelse(is.na(MNP_cell_data$Manual.Phenotype) & MNP_cell_data$seurat_clusters == 9, "B cell/Plasmablast", MNP_cell_data$Manual.Phenotype)
MNP_cell_data$Manual.Phenotype = ifelse(is.na(MNP_cell_data$Manual.Phenotype) & MNP_cell_data$seurat_clusters == 10, "B cell/Plasmablast", MNP_cell_data$Manual.Phenotype)
MNP_cell_data$Manual.Phenotype = ifelse(is.na(MNP_cell_data$Manual.Phenotype) & MNP_cell_data$seurat_clusters == 11, "Proximal tubule", MNP_cell_data$Manual.Phenotype)
MNP_cell_data$Manual.Phenotype = ifelse(is.na(MNP_cell_data$Manual.Phenotype) & MNP_cell_data$seurat_clusters == 12, "Intermediate Monocyte", MNP_cell_data$Manual.Phenotype)
```

```{r}
DimPlot(MNP_cell_data, reduction = "SCT_umap_mnp_cells", group.by = "Manual.Phenotype", label = T)
```

Number of "Macrophage" (annotated originally by SingleR as MNP) across the different slides:

```{r}
meta = MNP_cell_data@meta.data
meta %>% group_by(tissue, fov) %>% summarise(cells = n(), .groups = "keep") %>% spread(tissue, cells) %>% colSums(na.rm= T)
```

Lets add the manually annotated MNP labels back to the `manley_data`.

```{r, fig.height = 10, fig.width = 10}
manley_data$final_labels[manley_data$cell_ID %in% meta$cell_ID] = meta$Manual.Phenotype

DimPlot(manley_data, group.by = "final_labels", label = T, repel = T)  +
  theme(legend.position = 'none')
```

## Final Data Exporting and Phenotypes

```{r}
#saveRDS(manley_data, "Manley_SMI/results/rds/manley_data_tcell_mnp_manual_annotation.rds")
manley_data = readRDS("Manley_SMI/results/rds/manley_data_tcell_mnp_manual_annotation.rds")
clinical = read.csv("Manley_SMI/data/Clinical.data_acs.csv") %>%
  mutate(tissue = paste0("RCC", Slide),
         FOV = as.character(FOV))
```

```{r}
frequencies = manley_data@meta.data %>%
  group_by(tissue, fov, final_labels, annotation) %>%
  summarise(n = n()) %>%
  group_by(tissue, annotation) %>%
  group_map(~{
    .x %>% 
      spread(fov, n, fill = 0) %>%
      gather(fov, n, -final_labels) %>%
      mutate(annotation = .y$annotation,
             tissue = .y$tissue)
    }) %>%
  do.call(bind_rows, .) %>%
  group_by(tissue, fov) %>%
  mutate(frequency = n/sum(n)*100) %>%
  full_join(clinical, by = c("fov" = "FOV", "tissue" = "tissue"))
```

```{r, fig.width = 12, fig.height = 6}
frequencies %>%
    filter(tissue == "RCC3") %>%
    group_by(annotation) %>%
    group_map(~{
        plot_ly(.x, x = ~Specimen.ID.primary, y = ~frequency, color = ~final_labels, legendgroup = ~final_labels, type = "bar",showlegend = (.y == "Tumor ")) %>%
            layout(xaxis=list(
                title = paste0(unique(as.character(.y$annotation)))
            ), yaxis=list(title = "Frequency")) %>%
            layout(barmode = "stack")
    }) %>%
    subplot(nrows = 1, margin = 0.05, shareX=F, shareY =T, titleY= T, titleX=T) %>%
  layout(title = "RCC3 Tumor/Stroma")
```

```{r, fig.width = 12, fig.height = 6}
frequencies %>%
    filter(tissue == "RCC4") %>%
    group_by(annotation) %>%
    group_map(~{
        plot_ly(.x, x = ~Specimen.ID.primary, y = ~frequency, color = ~final_labels, legendgroup = ~final_labels, type = "bar",showlegend = (.y == "Tumor ")) %>%
            layout(xaxis=list(
                title = paste0(unique(as.character(.y$annotation)))
            ), yaxis=list(title = "Frequency")) %>%
            layout(barmode = "stack")
    }) %>%
    subplot(nrows = 1, margin = 0.05, shareX=F, shareY =T, titleY= T, titleX=T) %>%
  layout(title = "RCC4 Tumor/Stroma")
```

On RCC5, FOV 3 and 17 are missing from the seurat object.

```{r, fig.width = 12, fig.height = 6}
frequencies %>%
  filter(tissue == "RCC5") %>%
  group_by(annotation, fov) %>%
  mutate(allNA = !all(is.na(annotation))) %>%
  filter(allNA) %>%
  group_by(annotation) %>%
  group_map(~{
    plot_ly(.x, x = ~Specimen.ID.primary, y = ~frequency, color = ~final_labels, legendgroup = ~final_labels, type =
              "bar",showlegend = (.y == "Tumor ")) %>%
      layout(xaxis=list(
        title = paste0(unique(as.character(.y$annotation)))
        ), yaxis=list(title = "Frequency")) %>%
      layout(barmode = "stack")
    }) %>%
  subplot(nrows = 1, margin = 0.05, shareX=F, shareY =T, titleY= T, titleX=T) %>%
  layout(title = "RCC5 Tumor/Stroma")
```
